refactor portfolio workflow route
