+++ src/app/router_registry.py
@@ -124,6 +124,7 @@ from app.routers.portfolio_liquidity import router as portfolio_liquidity_router
 from app.routers.portfolio_performance import router as portfolio_performance_router
 from app.routers.portfolio_positions import router as portfolio_positions_router
 from app.routers.portfolio_transactions import router as portfolio_transactions_router
+from app.routers.portfolio_workflow import router as portfolio_workflow_router
 from app.routers.proposal_delivery import router as proposal_delivery_router
 from app.routers.proposal_memo_actions import router as proposal_memo_actions_router
 from app.routers.proposal_memo_evidence import router as proposal_memo_evidence_router
@@ -215,6 +216,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(portfolio_allocations_router)
     app.include_router(portfolio_positions_router)
     app.include_router(portfolio_performance_router)
+    app.include_router(portfolio_workflow_router)
     app.include_router(composite_performance_router)
     app.include_router(dpm_command_center_router)
     app.include_router(dpm_command_center_monitoring_router)
